+++ faynoSync_test.go
@@ -188,7 +188,8 @@ func TestSignUp(t *testing.T) {
 		handler.SignUp(c)
 	})
 
-	payload := `{"username": "admin", "password": "password", "api_key": "UHp3aKb40fwpoKZluZByWQ"}`
+	regKey := os.Getenv("API_KEY")
+	payload := fmt.Sprintf(`{"username": "admin", "password": "password", "api_key": "%s"}`, regKey)
 	req, err := http.NewRequest("POST", "/signup", bytes.NewBufferString(payload))
 	if err != nil {
 		t.Fatal(err)
@@ -299,13 +300,13 @@ func TestListApps(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /app/list endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.GET("/app/list", func(c *gin.Context) {
 		handler.ListApps(c)
 	})
 
-	// Create a POST request for the upload endpoint.
+	// Create a POST request for the /app/list endpoint.
 	req, err := http.NewRequest("GET", "/app/list", nil)
 	if err != nil {
 		t.Fatal(err)
@@ -402,7 +403,7 @@ func TestAppCreate(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the handler for the /channel/create route
+	// Define the handler for the /app/create route
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.POST("/app/create", func(c *gin.Context) {
 		handler.CreateApp(c)
@@ -429,7 +430,7 @@ func TestAppCreate(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	// Create a POST request to the /channel/create endpoint
+	// Create a POST request to the /app/create endpoint
 	req, err := http.NewRequest("POST", "/app/create", body)
 	if err != nil {
 		t.Fatal(err)
@@ -454,7 +455,7 @@ func TestAppCreate(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	// Check for the presence of the "createChannelResult.Created" key in the response
+	// Check for the presence of the "createAppResult.Created" key in the response
 	id, idExists := response["createAppResult.Created"]
 	assert.True(t, idExists)
 	idTestappApp = id.(string)
@@ -468,7 +469,7 @@ func TestSecondaryAppCreate(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the handler for the /channel/create route
+	// Define the handler for the /app/create route
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.POST("/app/create", func(c *gin.Context) {
 		handler.CreateApp(c)
@@ -495,7 +496,7 @@ func TestSecondaryAppCreate(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	// Create a POST request to the /channel/create endpoint
+	// Create a POST request to the /app/create endpoint
 	req, err := http.NewRequest("POST", "/app/create", body)
 	if err != nil {
 		t.Fatal(err)
@@ -524,7 +525,7 @@ func TestUpload(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /upload endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.POST("/upload", func(c *gin.Context) {
 		handler.UploadApp(c)
@@ -565,7 +566,7 @@ func TestUpload(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	// Create a POST request for the upload endpoint.
+	// Create a POST request for the /upload endpoint.
 	req, err := http.NewRequest("POST", "/upload", body)
 	if err != nil {
 		t.Fatal(err)
@@ -587,6 +588,7 @@ func TestUpload(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
+	logrus.Infof("Response Body after unmarshaling: %+v", response)
 	// Check that the "token" key exists in the response.
 	id, idExists := response["uploadResult.Uploaded"]
 	assert.True(t, idExists)
@@ -602,7 +604,7 @@ func TestUploadDuplicateApp(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /upload endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.POST("/upload", func(c *gin.Context) {
 		handler.UploadApp(c)
@@ -643,7 +645,7 @@ func TestUploadDuplicateApp(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	// Create a POST request for the upload endpoint.
+	// Create a POST request for the /upload endpoint.
 	req, err := http.NewRequest("POST", "/upload", body)
 	if err != nil {
 		t.Fatal(err)
@@ -661,7 +663,7 @@ func TestUploadDuplicateApp(t *testing.T) {
 	assert.Equal(t, http.StatusInternalServerError, w.Code)
 
 	// Check the response body for the desired error message.
-	expectedErrorMessage := `{"error":"app with this name, version, and extension already exists"}`
+	expectedErrorMessage := `{"error":"app with this name, version, platform, architecture and extension already exists"}`
 	assert.Equal(t, expectedErrorMessage, w.Body.String())
 }
 
@@ -671,13 +673,13 @@ func TestDeleteApp(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /apps/delete endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.DELETE("/apps/delete", func(c *gin.Context) {
 		handler.DeleteSpecificVersionOfApp(c)
 	})
 
-	// Create a POST request for the upload endpoint.
+	// Create a POST request for the /apps/delete endpoint.
 	req, err := http.NewRequest("DELETE", "/apps/delete?id="+uploadedFirstApp, nil)
 	if err != nil {
 		t.Fatal(err)
@@ -701,13 +703,13 @@ func TestListChannels(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /channel/list endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.GET("/channel/list", func(c *gin.Context) {
 		handler.ListChannels(c)
 	})
 
-	// Create a POST request for the upload endpoint.
+	// Create a POST request for the /channel/list endpoint.
 	req, err := http.NewRequest("GET", "/channel/list", nil)
 	if err != nil {
 		t.Fatal(err)
@@ -793,7 +795,59 @@ func TestChannelCreateNightly(t *testing.T) {
 	assert.True(t, idExists)
 	assert.NotEmpty(t, idNightlyChannel)
 }
+func TestChannelCreateWithWrongName(t *testing.T) {
+	// Initialize Gin router and recorder for the test
+	router := gin.Default()
+	router.Use(utils.AuthMiddleware())
+	w := httptest.NewRecorder()
 
+	// Define the handler for the /channel/create route
+	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
+	router.POST("/channel/create", func(c *gin.Context) {
+		handler.CreateChannel(c)
+	})
+
+	// Create multipart/form-data request body
+	body := &bytes.Buffer{}
+	writer := multipart.NewWriter(body)
+
+	// Add a field for the channel to the form
+	dataPart, err := writer.CreateFormField("data")
+	if err != nil {
+		t.Fatal(err)
+	}
+	payload := `{"channel": "nightly *"}`
+	_, err = dataPart.Write([]byte(payload))
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Close the writer to finalize the form data
+	err = writer.Close()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Create a POST request to the /channel/create endpoint
+	req, err := http.NewRequest("POST", "/channel/create", body)
+	if err != nil {
+		t.Fatal(err)
+	}
+	// Set the Authorization header.
+	req.Header.Set("Authorization", "Bearer "+authToken)
+	// Set the Content-Type header for multipart/form-data
+	req.Header.Set("Content-Type", writer.FormDataContentType())
+
+	// Serve the request using the Gin router
+	router.ServeHTTP(w, req)
+	logrus.Infoln("Response Body:", w.Body.String())
+	// Check the response status code (expecting 400).
+	assert.Equal(t, http.StatusBadRequest, w.Code)
+
+	// Check the response body for the desired error message.
+	expectedErrorMessage := `{"error":"invalid channel name"}`
+	assert.Equal(t, expectedErrorMessage, w.Body.String())
+}
 func TestSecondaryChannelCreateNightly(t *testing.T) {
 	// Initialize Gin router and recorder for the test
 	router := gin.Default()
@@ -920,7 +974,7 @@ func TestUploadAppWithoutChannel(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /upload endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.POST("/upload", func(c *gin.Context) {
 		handler.UploadApp(c)
@@ -989,13 +1043,13 @@ func TestListPlatforms(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /platform/list endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.GET("/platform/list", func(c *gin.Context) {
 		handler.ListPlatforms(c)
 	})
 
-	// Create a POST request for the upload endpoint.
+	// Create a POST request for the /platform/list endpoint.
 	req, err := http.NewRequest("GET", "/platform/list", nil)
 	if err != nil {
 		t.Fatal(err)
@@ -1047,7 +1101,7 @@ func TestPlatformCreate(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	// Create a POST request to the /channel/create endpoint
+	// Create a POST request to the /platform/create endpoint
 	req, err := http.NewRequest("POST", "/platform/create", body)
 	if err != nil {
 		t.Fatal(err)
@@ -1071,7 +1125,6 @@ func TestPlatformCreate(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	// Check that the "token" key exists in the response.
 	id, idExists := response["createPlatformResult.Created"]
 	assert.True(t, idExists)
 	platformId = id.(string)
@@ -1109,7 +1162,7 @@ func TestSecondaryPlatformCreate(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	// Create a POST request to the /channel/create endpoint
+	// Create a POST request to the /platform/create endpoint
 	req, err := http.NewRequest("POST", "/platform/create", body)
 	if err != nil {
 		t.Fatal(err)
@@ -1135,7 +1188,7 @@ func TestUploadAppWithoutPlatform(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /upload endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.POST("/upload", func(c *gin.Context) {
 		handler.UploadApp(c)
@@ -1204,13 +1257,13 @@ func TestListArchs(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /arch/list endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.GET("/arch/list", func(c *gin.Context) {
 		handler.ListArchs(c)
 	})
 
-	// Create a POST request for the upload endpoint.
+	// Create a POST request for the /arch/list endpoint.
 	req, err := http.NewRequest("GET", "/arch/list", nil)
 	if err != nil {
 		t.Fatal(err)
@@ -1287,7 +1340,6 @@ func TestArchCreate(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	// Check that the "token" key exists in the response.
 	id, idExists := response["createArchResult.Created"]
 	assert.True(t, idExists)
 	archId = id.(string)
@@ -1308,7 +1360,7 @@ func TestSecondaryArchCreate(t *testing.T) {
 	body := &bytes.Buffer{}
 	writer := multipart.NewWriter(body)
 
-	// Add a field for the channel to the form
+	// Add a field for the arch to the form
 	dataPart, err := writer.CreateFormField("data")
 	if err != nil {
 		t.Fatal(err)
@@ -1351,7 +1403,7 @@ func TestUploadAppWithoutArch(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /upload endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.POST("/upload", func(c *gin.Context) {
 		handler.UploadApp(c)
@@ -1428,7 +1480,7 @@ func TestMultipleUpload(t *testing.T) {
 	})
 
 	// Create a file to upload (you can replace this with a test file path).
-	filePaths := []string{"testapp.dmg", "testapp.pkg"}
+	filePaths := []string{"testapp.dmg", "testapp.pkg", "LICENSE"}
 	for _, filePath := range filePaths {
 		file, err := os.Open(filePath)
 		if err != nil {
@@ -1480,7 +1532,7 @@ func TestMultipleUpload(t *testing.T) {
 			if err != nil {
 				t.Fatal(err)
 			}
-			// Create a POST request for the upload endpoint with the current combination.
+			// Create a POST request for the /upload endpoint with the current combination.
 			req, err := http.NewRequest("POST", "/upload", body)
 			if err != nil {
 				t.Fatal(err)
@@ -1502,7 +1554,6 @@ func TestMultipleUpload(t *testing.T) {
 				t.Fatal(err)
 			}
 
-			// Check that the "token" key exists in the response.
 			id, idExists := response["uploadResult.Uploaded"]
 			assert.True(t, idExists)
 
@@ -1530,13 +1581,13 @@ func TestUpdateSpecificApp(t *testing.T) {
 
 	router := gin.Default()
 	router.Use(utils.AuthMiddleware())
-	// Define the route for the upload endpoint.
+	// Define the route for the update endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.POST("/apps/update", func(c *gin.Context) {
 		handler.UpdateSpecificApp(c)
 	})
 
-	// Create a file to upload (you can replace this with a test file path).
+	// Create a file to update (you can replace this with a test file path).
 	filePaths := []string{"LICENSE", "LICENSE"}
 	for _, filePath := range filePaths {
 		file, err := os.Open(filePath)
@@ -1558,7 +1609,7 @@ func TestUpdateSpecificApp(t *testing.T) {
 			{uploadedAppIDs[1], "0.0.2.137", "nightly", true, true, "universalPlatform", "universalArch", "### Changelog"},
 		}
 
-		// Iterate through the combinations and upload the file for each combination.
+		// Iterate through the combinations and update the file for each combination.
 		for _, combo := range combinations {
 			w := httptest.NewRecorder()
 			// Reset the request body for each iteration.
@@ -1572,7 +1623,7 @@ func TestUpdateSpecificApp(t *testing.T) {
 			if err != nil {
 				t.Fatal(err)
 			}
-			// Create a POST request for the upload endpoint with the current combination.
+			// Create a POST request for the update endpoint with the current combination.
 			dataPart, err := writer.CreateFormField("data")
 			if err != nil {
 				t.Fatal(err)
@@ -1619,13 +1670,13 @@ func TestSearch(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /search endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.GET("/search", func(c *gin.Context) {
 		handler.GetAppByName(c)
 	})
 
-	// Create a POST request for the upload endpoint.
+	// Create a POST request for the /search endpoint.
 	req, err := http.NewRequest("GET", "/search?app_name=testapp", nil)
 	if err != nil {
 		t.Fatal(err)
@@ -1657,7 +1708,6 @@ func TestSearch(t *testing.T) {
 
 	expected := []AppInfo{
 		{
-			// AppID:     idTestappApp,
 			AppName:   "testapp",
 			Version:   "0.0.1.137",
 			Channel:   "nightly",
@@ -1674,6 +1724,11 @@ func TestSearch(t *testing.T) {
 					Arch:     "universalArch",
 					Package:  ".pkg",
 				},
+				{
+					Platform: "universalPlatform",
+					Arch:     "universalArch",
+					Package:  "",
+				},
 			},
 			Changelog: []model.Changelog{
 				{
@@ -1684,7 +1739,6 @@ func TestSearch(t *testing.T) {
 			},
 		},
 		{
-			// AppID:     idTestappApp,
 			AppName:   "testapp",
 			Version:   "0.0.2.137",
 			Channel:   "nightly",
@@ -1701,6 +1755,11 @@ func TestSearch(t *testing.T) {
 					Arch:     "universalArch",
 					Package:  ".pkg",
 				},
+				{
+					Platform: "universalPlatform",
+					Arch:     "universalArch",
+					Package:  "",
+				},
 			},
 			Changelog: []model.Changelog{
 				{
@@ -1711,7 +1770,6 @@ func TestSearch(t *testing.T) {
 			},
 		},
 		{
-			// AppID:     idTestappApp,
 			AppName:   "testapp",
 			Version:   "0.0.3.137",
 			Channel:   "nightly",
@@ -1728,6 +1786,11 @@ func TestSearch(t *testing.T) {
 					Arch:     "universalArch",
 					Package:  ".pkg",
 				},
+				{
+					Platform: "universalPlatform",
+					Arch:     "universalArch",
+					Package:  "",
+				},
 			},
 			Changelog: []model.Changelog{
 				{
@@ -1738,7 +1801,6 @@ func TestSearch(t *testing.T) {
 			},
 		},
 		{
-			// AppID:     idTestappApp,
 			AppName:   "testapp",
 			Version:   "0.0.4.137",
 			Channel:   "stable",
@@ -1755,6 +1817,11 @@ func TestSearch(t *testing.T) {
 					Arch:     "universalArch",
 					Package:  ".pkg",
 				},
+				{
+					Platform: "universalPlatform",
+					Arch:     "universalArch",
+					Package:  "",
+				},
 			},
 			Changelog: []model.Changelog{
 				{
@@ -1765,7 +1832,6 @@ func TestSearch(t *testing.T) {
 			},
 		},
 		{
-			// AppID:     idTestappApp,
 			AppName:   "testapp",
 			Version:   "0.0.5.137",
 			Channel:   "stable",
@@ -1781,6 +1847,11 @@ func TestSearch(t *testing.T) {
 					Platform: "universalPlatform",
 					Arch:     "universalArch",
 					Package:  ".pkg",
+				},
+				{
+					Platform: "universalPlatform",
+					Arch:     "universalArch",
+					Package:  "",
 				},
 			},
 			Changelog: []model.Changelog{
@@ -1806,7 +1877,6 @@ func TestSearch(t *testing.T) {
 	}
 
 	for i, expectedApp := range expected {
-		// assert.Equal(t, expectedApp.AppID, actual.Apps[i].AppID)
 		assert.Equal(t, expectedApp.AppName, actual.Apps[i].AppName)
 		assert.Equal(t, expectedApp.Version, actual.Apps[i].Version)
 		assert.Equal(t, expectedApp.Channel, actual.Apps[i].Channel)
@@ -1861,6 +1931,9 @@ func TestFetchkLatestVersionOfApp(t *testing.T) {
 							"pkg": map[string]interface{}{
 								"url": fmt.Sprintf("%s/%s", s3Endpoint, url.PathEscape("testapp/nightly/universalPlatform/universalArch/testapp-0.0.2.137.pkg")),
 							},
+							"no-extension": map[string]interface{}{
+								"url": fmt.Sprintf("%s/%s", s3Endpoint, url.PathEscape("testapp/nightly/universalPlatform/universalArch/testapp-0.0.2.137")),
+							},
 						},
 					},
 				},
@@ -1882,6 +1955,9 @@ func TestFetchkLatestVersionOfApp(t *testing.T) {
 							},
 							"pkg": map[string]interface{}{
 								"url": fmt.Sprintf("%s/%s", s3Endpoint, url.PathEscape("testapp/stable/universalPlatform/universalArch/testapp-0.0.4.137.pkg")),
+							},
+							"no-extension": map[string]interface{}{
+								"url": fmt.Sprintf("%s/%s", s3Endpoint, url.PathEscape("testapp/stable/universalPlatform/universalArch/testapp-0.0.4.137")),
 							},
 						},
 					},
@@ -1949,12 +2025,12 @@ func TestCheckVersion(t *testing.T) {
 				"critical":         true,
 				"update_url_dmg":   fmt.Sprintf("%s/%s", s3Endpoint, url.PathEscape("testapp/nightly/universalPlatform/universalArch/testapp-0.0.2.137.dmg")),
 				"update_url_pkg":   fmt.Sprintf("%s/%s", s3Endpoint, url.PathEscape("testapp/nightly/universalPlatform/universalArch/testapp-0.0.2.137.pkg")),
+				"update_url":       fmt.Sprintf("%s/%s", s3Endpoint, url.PathEscape("testapp/nightly/universalPlatform/universalArch/testapp-0.0.2.137")),
 			},
 			ExpectedCode: http.StatusOK,
-			// Published:    false,
-			Platform: "universalPlatform",
-			Arch:     "universalArch",
-			TestName: "NightlyUpdateAvailable",
+			Platform:     "universalPlatform",
+			Arch:         "universalArch",
+			TestName:     "NightlyUpdateAvailable",
 		},
 		{
 			AppName:     "testapp",
@@ -1964,12 +2040,12 @@ func TestCheckVersion(t *testing.T) {
 				"update_available": false,
 				"update_url_dmg":   fmt.Sprintf("%s/%s", s3Endpoint, url.PathEscape("testapp/nightly/universalPlatform/universalArch/testapp-0.0.2.137.dmg")),
 				"update_url_pkg":   fmt.Sprintf("%s/%s", s3Endpoint, url.PathEscape("testapp/nightly/universalPlatform/universalArch/testapp-0.0.2.137.pkg")),
+				"update_url":       fmt.Sprintf("%s/%s", s3Endpoint, url.PathEscape("testapp/nightly/universalPlatform/universalArch/testapp-0.0.2.137")),
 			},
 			ExpectedCode: http.StatusOK,
-			// Published:    true,
-			Platform: "universalPlatform",
-			Arch:     "universalArch",
-			TestName: "NightlyUpdateAvailable",
+			Platform:     "universalPlatform",
+			Arch:         "universalArch",
+			TestName:     "NightlyUpdateAvailable",
 		},
 		{
 			AppName:     "testapp",
@@ -1979,10 +2055,9 @@ func TestCheckVersion(t *testing.T) {
 				"error": "requested version 0.0.3.137 is newer than the latest version available",
 			},
 			ExpectedCode: http.StatusInternalServerError,
-			// Published:    false,
-			Platform: "universalPlatform",
-			Arch:     "universalArch",
-			TestName: "NightlyUpdateAvailable",
+			Platform:     "universalPlatform",
+			Arch:         "universalArch",
+			TestName:     "NightlyUpdateAvailable",
 		},
 		{
 			AppName:     "testapp",
@@ -1992,12 +2067,12 @@ func TestCheckVersion(t *testing.T) {
 				"update_available": false,
 				"update_url_dmg":   fmt.Sprintf("%s/%s", s3Endpoint, url.PathEscape("testapp/stable/universalPlatform/universalArch/testapp-0.0.4.137.dmg")),
 				"update_url_pkg":   fmt.Sprintf("%s/%s", s3Endpoint, url.PathEscape("testapp/stable/universalPlatform/universalArch/testapp-0.0.4.137.pkg")),
+				"update_url":       fmt.Sprintf("%s/%s", s3Endpoint, url.PathEscape("testapp/stable/universalPlatform/universalArch/testapp-0.0.4.137")),
 			},
 			ExpectedCode: http.StatusOK,
-			// Published:    true,
-			Platform: "universalPlatform",
-			Arch:     "universalArch",
-			TestName: "StableUpdateAvailable",
+			Platform:     "universalPlatform",
+			Arch:         "universalArch",
+			TestName:     "StableUpdateAvailable",
 		},
 		{
 			AppName:     "testapp",
@@ -2047,7 +2122,7 @@ func TestMultipleDelete(t *testing.T) {
 	router := gin.Default()
 	router.Use(utils.AuthMiddleware())
 
-	// Define the route for the deleteApp endpoint.
+	// Define the route for the /apps/delete endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.DELETE("/apps/delete", func(c *gin.Context) {
 		handler.DeleteSpecificVersionOfApp(c)
@@ -2143,13 +2218,13 @@ func TestListChannelsWhenExist(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /channel/list endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.GET("/channel/list", func(c *gin.Context) {
 		handler.ListChannels(c)
 	})
 
-	// Create a POST request for the upload endpoint.
+	// Create a POST request for the /channel/list endpoint.
 	req, err := http.NewRequest("GET", "/channel/list", nil)
 	if err != nil {
 		t.Fatal(err)
@@ -2197,13 +2272,13 @@ func TestDeleteNightlyChannel(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /channel/delete endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.DELETE("/channel/delete", func(c *gin.Context) {
 		handler.DeleteChannel(c)
 	})
 
-	// Create a POST request for the upload endpoint.
+	// Create a DELETE request for the /channel/delete endpoint.
 	req, err := http.NewRequest("DELETE", "/channel/delete?id="+idNightlyChannel, nil)
 	if err != nil {
 		t.Fatal(err)
@@ -2227,13 +2302,13 @@ func TestDeleteStableChannel(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /channel/delete endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.DELETE("/channel/delete", func(c *gin.Context) {
 		handler.DeleteChannel(c)
 	})
 
-	// Create a POST request for the upload endpoint.
+	// Create a DELETE request for the /channel/delete endpoint.
 	req, err := http.NewRequest("DELETE", "/channel/delete?id="+idStableChannel, nil)
 	if err != nil {
 		t.Fatal(err)
@@ -2256,7 +2331,7 @@ func TestUpdatePlatform(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the handler for the /UpdatePlatform route
+	// Define the handler for the /platform/update route
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.POST("/platform/update", func(c *gin.Context) {
 		handler.UpdatePlatform(c)
@@ -2313,19 +2388,74 @@ func TestUpdatePlatform(t *testing.T) {
 	assert.True(t, exists)
 	assert.True(t, updated.(bool))
 }
+
+func TestFailedUpdatePlatform(t *testing.T) {
+	// Initialize Gin router and recorder for the test
+	router := gin.Default()
+	router.Use(utils.AuthMiddleware())
+	w := httptest.NewRecorder()
+
+	// Define the handler for the /platform/update route
+	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
+	router.POST("/platform/update", func(c *gin.Context) {
+		handler.UpdatePlatform(c)
+	})
+
+	// Create multipart/form-data request body
+	body := &bytes.Buffer{}
+	writer := multipart.NewWriter(body)
+
+	// Add a field for the channel to the form
+	dataPart, err := writer.CreateFormField("data")
+	if err != nil {
+		t.Fatal(err)
+	}
+	payload := fmt.Sprintf(`{"id": "%s", "platform":"new * Platform"}`, platformId)
+	_, err = dataPart.Write([]byte(payload))
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Close the writer to finalize the form data
+	err = writer.Close()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Create a POST request to the /platform/update endpoint
+	req, err := http.NewRequest("POST", "/platform/update", body)
+	if err != nil {
+		t.Fatal(err)
+	}
+	// Set the Authorization header.
+	req.Header.Set("Authorization", "Bearer "+authToken)
+	// Set the Content-Type header for multipart/form-data
+	req.Header.Set("Content-Type", writer.FormDataContentType())
+
+	// Serve the request using the Gin router
+	router.ServeHTTP(w, req)
+	logrus.Infoln("Response Body:", w.Body.String())
+	// Check the response status code (expecting 400).
+	assert.Equal(t, http.StatusBadRequest, w.Code)
+
+	// Check the response body for the desired error message.
+	expectedErrorMessage := `{"error":"invalid platform name"}`
+	assert.Equal(t, expectedErrorMessage, w.Body.String())
+}
+
 func TestListPlatformsWhenExist(t *testing.T) {
 
 	router := gin.Default()
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /platform/list endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.GET("/platform/list", func(c *gin.Context) {
 		handler.ListPlatforms(c)
 	})
 
-	// Create a POST request for the upload endpoint.
+	// Create a POST request for the /platform/list endpoint.
 	req, err := http.NewRequest("GET", "/platform/list", nil)
 	if err != nil {
 		t.Fatal(err)
@@ -2370,13 +2500,13 @@ func TestDeletePlatform(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /platform/delete endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.DELETE("/platform/delete", func(c *gin.Context) {
 		handler.DeletePlatform(c)
 	})
 
-	// Create a POST request for the upload endpoint.
+	// Create a DELETE request for the /platform/delete endpoint.
 	req, err := http.NewRequest("DELETE", "/platform/delete?id="+platformId, nil)
 	if err != nil {
 		t.Fatal(err)
@@ -2462,13 +2592,13 @@ func TestListArchsWhenExist(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /arch/list endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.GET("/arch/list", func(c *gin.Context) {
 		handler.ListArchs(c)
 	})
 
-	// Create a POST request for the upload endpoint.
+	// Create a POST request for the /arch/list endpoint.
 	req, err := http.NewRequest("GET", "/arch/list", nil)
 	if err != nil {
 		t.Fatal(err)
@@ -2513,13 +2643,13 @@ func TestDeleteArch(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /arch/delete endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.DELETE("/arch/delete", func(c *gin.Context) {
 		handler.DeleteArch(c)
 	})
 
-	// Create a POST request for the upload endpoint.
+	// Create a DELETE request for the /arch/delete endpoint.
 	req, err := http.NewRequest("DELETE", "/arch/delete?id="+archId, nil)
 	if err != nil {
 		t.Fatal(err)
@@ -2594,7 +2724,7 @@ func TestUpdateApp(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	// Check for the presence of the "updateChannelResult.Updated" key in the response
+	// Check for the presence of the "updateAppResult.Updated" key in the response
 	updated, exists := response["updateAppResult.Updated"]
 	assert.True(t, exists)
 	assert.True(t, updated.(bool))
@@ -2605,13 +2735,13 @@ func TestListAppsWhenExist(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /app/list endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.GET("/app/list", func(c *gin.Context) {
 		handler.ListApps(c)
 	})
 
-	// Create a POST request for the upload endpoint.
+	// Create a GET request for the /app/list endpoint.
 	req, err := http.NewRequest("GET", "/app/list", nil)
 	if err != nil {
 		t.Fatal(err)
@@ -2656,13 +2786,13 @@ func TestDeleteAppMeta(t *testing.T) {
 	router.Use(utils.AuthMiddleware())
 	w := httptest.NewRecorder()
 
-	// Define the route for the upload endpoint.
+	// Define the route for the /app/delete endpoint.
 	handler := handler.NewAppHandler(client, appDB, mongoDatabase, redisClient, true)
 	router.DELETE("/app/delete", func(c *gin.Context) {
 		handler.DeleteApp(c)
 	})
 
-	// Create a POST request for the upload endpoint.
+	// Create a DELETE request for the /app/delete endpoint.
 	req, err := http.NewRequest("DELETE", "/app/delete?id="+idTestappApp, nil)
 	if err != nil {
 		t.Fatal(err)